ImagesTable: Split ImagesTable index into component files
The original file included both the table and row components, this splits them into smaller individual files to improve readability.
