+++ src/Components/ImagesTable/components/Row/index.ts
@@ -1,0 +1,9 @@
+export { default as AwsRow } from './components/AwsRow';
+export { default as AwsS3Row } from './components/AwsS3Row';
+export { default as AzureRow } from './components/AzureRow';
+export { default as GcpRow } from './components/GcpRow';
+export { default as ImagesTableRow } from './components/ImagesTableRow';
+export { default as LocalRow } from './components/LocalRow';
+export { default as OciRow } from './components/OciRow';
+export { default as Row } from './components/Row';
+export { defaultActions } from './defaultActions';
