+++ src/Components/ImagesTable/index.tsx
@@ -1,13 +1,8 @@
-import React, { useEffect, useRef, useState } from 'react';
-
-import path from 'path';
+import React, { useState } from 'react';
 
 import {
   Alert,
-  Badge,
   Bullseye,
-  Button,
-  Label,
   PageSection,
   Pagination,
   PaginationVariant,
@@ -17,30 +12,16 @@ import {
   ToolbarItem,
 } from '@patternfly/react-core';
 import { OnSetPage } from '@patternfly/react-core/dist/esm/components/Pagination/Pagination';
-import {
-  ActionsColumn,
-  ExpandableRowContent,
-  Table,
-  Tbody,
-  Td,
-  Th,
-  Thead,
-  Tr,
-} from '@patternfly/react-table';
+import { Table, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table';
 import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
-import cockpit from 'cockpit';
-import { useDispatch } from 'react-redux';
 
 import {
   BlueprintItem,
-  ComposesResponseItem,
   GetBlueprintComposesApiArg,
   GetBlueprintsApiArg,
-  LocalUploadStatus,
   useGetBlueprintComposesQuery,
   useGetBlueprintsQuery,
   useGetComposesQuery,
-  useGetComposeStatusQuery,
 } from '@/store/api/backend';
 import {
   selectBlueprintSearchInput,
@@ -48,52 +29,20 @@ import {
   selectBlueprintVersionFilterAPI,
   selectLimit,
   selectOffset,
-  selectSelectedBlueprintId,
-  setBlueprintId,
 } from '@/store/slices/blueprint';
 import { selectIsOnPremise } from '@/store/slices/env';
 
 import ImagesEmptyState from './components/EmptyState';
-import {
-  AwsDetails,
-  AwsS3Details,
-  AzureDetails,
-  GcpDetails,
-  LocalDetails,
-  OciDetails,
-} from './components/ImageDetails';
 import ImagesTableToolbar from './components/ImagesTableToolbar';
-import { AwsS3Instance, LocalInstance } from './components/Instance';
-import Release from './components/Release';
-import { CloudStatus, ExpiringStatus, LocalStatus } from './components/Status';
-import { AwsTarget, Target } from './components/Target';
+import { ImagesTableRow } from './components/Row';
 
 import {
-  AMPLITUDE_MODULE_NAME,
-  AWS_S3_EXPIRATION_TIME_IN_HOURS,
-  OCI_STORAGE_EXPIRATION_TIME_IN_DAYS,
   PAGINATION_LIMIT,
   PAGINATION_OFFSET,
   SEARCH_INPUT,
-  STATUS_POLLING_INTERVAL,
 } from '../../constants';
-import {
-  useCockpitMachinesAvailable,
-  useEffectiveBlueprintId,
-  useGetUser,
-} from '../../Hooks';
+import { useEffectiveBlueprintId, useGetUser } from '../../Hooks';
 import { useAppSelector } from '../../store/hooks';
-import { hasBootcRequest } from '../../store/typeGuards';
-import { bootcReferenceToOSDisplayLabel } from '../../Utilities/distributionToOSShortId';
-import {
-  computeHoursToExpiration,
-  timestampToDisplayString,
-  timestampToDisplayStringDetailed,
-} from '../../Utilities/time';
-import { AWSLaunchModal } from '../Launch/AWSLaunchModal';
-import { AzureLaunchModal } from '../Launch/AzureLaunchModal';
-import { GcpLaunchModal } from '../Launch/GcpLaunchModal';
-import { OciLaunchModal } from '../Launch/OciLaunchModal';
 
 const ImagesTable = () => {
   const [page, setPage] = useState(1);
@@ -314,454 +263,6 @@ const ImagesTable = () => {
       </Toolbar>
     </PageSection>
   );
-};
-
-type ImagesTableRowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const ImagesTableRow = ({ compose, rowIndex }: ImagesTableRowPropTypes) => {
-  const [pollingInterval, setPollingInterval] = useState(
-    STATUS_POLLING_INTERVAL,
-  );
-  const lastTrackedStatusRef = useRef<string | null>(null);
-  const { analytics, auth } = useChrome();
-  const { userData } = useGetUser(auth);
-  const isOnPremise = useAppSelector(selectIsOnPremise);
-
-  const { data: composeStatus } = useGetComposeStatusQuery(
-    {
-      composeId: compose.id,
-    },
-    { pollingInterval: pollingInterval },
-  );
-
-  useEffect(() => {
-    if (
-      composeStatus?.image_status.status === 'success' ||
-      composeStatus?.image_status.status === 'failure'
-    ) {
-      setPollingInterval(0);
-    } else {
-      setPollingInterval(STATUS_POLLING_INTERVAL);
-    }
-  }, [setPollingInterval, composeStatus]);
-
-  useEffect(() => {
-    const currentStatus = composeStatus?.image_status.status;
-    if (!isOnPremise && currentStatus) {
-      if (lastTrackedStatusRef.current === null) {
-        lastTrackedStatusRef.current = currentStatus;
-        return;
-      }
-
-      const buildIsCompleted =
-        currentStatus === 'success' || currentStatus === 'failure';
-      const statusChanged = lastTrackedStatusRef.current !== currentStatus;
-
-      if (buildIsCompleted && statusChanged) {
-        lastTrackedStatusRef.current = currentStatus;
-        const imageType = compose.request.image_requests[0]?.image_type;
-        const uploadType =
-          compose.request.image_requests[0]?.upload_request?.type;
-        const isError = currentStatus === 'failure';
-        const error = composeStatus.image_status.error;
-        analytics.track(
-          `${AMPLITUDE_MODULE_NAME} - Image Creation - ${isError ? 'Failure' : 'Success'}`,
-          {
-            module: AMPLITUDE_MODULE_NAME,
-            error: isError,
-            image_type: imageType,
-            upload_type: uploadType,
-            compose_id: compose.id,
-            ...(isError
-              ? {
-                  error_id: error?.id,
-                  error_details: error?.details,
-                  error_reason: error?.reason,
-                }
-              : {}),
-            account_id: userData?.identity.internal?.account_id || 'Not found',
-          },
-        );
-      } else if (statusChanged) {
-        lastTrackedStatusRef.current = currentStatus;
-      }
-    }
-  }, [analytics, userData, compose, composeStatus]);
-
-  const type = compose.request.image_requests[0]?.upload_request?.type;
-
-  switch (type as string) {
-    case 'aws':
-      return <AwsRow compose={compose} rowIndex={rowIndex} />;
-    case 'gcp':
-      return <GcpRow compose={compose} rowIndex={rowIndex} />;
-    case 'azure':
-      return <AzureRow compose={compose} rowIndex={rowIndex} />;
-    case 'oci.objectstorage':
-      return <OciRow compose={compose} rowIndex={rowIndex} />;
-    case 'aws.s3':
-      return <AwsS3Row compose={compose} rowIndex={rowIndex} />;
-    case 'local':
-      return <LocalRow compose={compose} rowIndex={rowIndex} />;
-  }
-};
-
-type GcpRowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const GcpRow = ({ compose, rowIndex }: GcpRowPropTypes) => {
-  const details = <GcpDetails compose={compose} />;
-  const instance = <GcpLaunchModal compose={compose} />;
-  const status = <CloudStatus compose={compose} />;
-
-  return (
-    <Row
-      compose={compose}
-      rowIndex={rowIndex}
-      details={details}
-      status={status}
-      instance={instance}
-    />
-  );
-};
-
-type AzureRowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const AzureRow = ({ compose, rowIndex }: AzureRowPropTypes) => {
-  const details = <AzureDetails compose={compose} />;
-  const instance = <AzureLaunchModal compose={compose} />;
-  const status = <CloudStatus compose={compose} />;
-
-  return (
-    <Row
-      compose={compose}
-      rowIndex={rowIndex}
-      details={details}
-      instance={instance}
-      status={status}
-    />
-  );
-};
-
-type OciRowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const OciRow = ({ compose, rowIndex }: OciRowPropTypes) => {
-  const daysToExpiration = Math.floor(
-    computeHoursToExpiration(compose.created_at) / 24,
-  );
-  const isExpired = daysToExpiration >= OCI_STORAGE_EXPIRATION_TIME_IN_DAYS;
-
-  const details = <OciDetails compose={compose} />;
-  const instance = <OciLaunchModal compose={compose} isExpired={isExpired} />;
-  const status = (
-    <ExpiringStatus
-      compose={compose}
-      isExpired={isExpired}
-      timeToExpiration={daysToExpiration}
-    />
-  );
-
-  return (
-    <Row
-      compose={compose}
-      rowIndex={rowIndex}
-      details={details}
-      instance={instance}
-      status={status}
-    />
-  );
-};
-
-type AwsS3RowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const AwsS3Row = ({ compose, rowIndex }: AwsS3RowPropTypes) => {
-  const hoursToExpiration = computeHoursToExpiration(compose.created_at);
-  const awsS3ExpirationTime = AWS_S3_EXPIRATION_TIME_IN_HOURS;
-  const isExpired = hoursToExpiration >= awsS3ExpirationTime;
-
-  const details = <AwsS3Details compose={compose} />;
-  const instance = <AwsS3Instance compose={compose} isExpired={isExpired} />;
-  const status = (
-    <ExpiringStatus
-      compose={compose}
-      isExpired={isExpired}
-      timeToExpiration={hoursToExpiration}
-    />
-  );
-
-  return (
-    <Row
-      compose={compose}
-      rowIndex={rowIndex}
-      details={details}
-      instance={instance}
-      status={status}
-    />
-  );
-};
-
-type AwsRowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const AwsRow = ({ compose, rowIndex }: AwsRowPropTypes) => {
-  const target = <AwsTarget />;
-  const status = <CloudStatus compose={compose} />;
-  const instance = <AWSLaunchModal compose={compose} />;
-  const details = <AwsDetails compose={compose} />;
-
-  return (
-    <Row
-      compose={compose}
-      rowIndex={rowIndex}
-      status={status}
-      target={target}
-      instance={instance}
-      details={details}
-    />
-  );
-};
-
-type LocalRowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-};
-
-const LocalRow = ({ compose, rowIndex }: LocalRowPropTypes) => {
-  const details = <LocalDetails compose={compose} />;
-  const instance = <LocalInstance compose={compose} />;
-  const status = <LocalStatus compose={compose} />;
-  return (
-    <Row
-      compose={compose}
-      rowIndex={rowIndex}
-      details={details}
-      instance={instance}
-      status={status}
-    />
-  );
-};
-
-type RowPropTypes = {
-  compose: ComposesResponseItem;
-  rowIndex: number;
-  status: JSX.Element;
-  target?: JSX.Element;
-  actions?: JSX.Element;
-  instance: JSX.Element;
-  details: JSX.Element;
-};
-
-type Analytics = {
-  track: (event: string, props?: Record<string, unknown>) => void;
-};
-
-const Row = ({
-  compose,
-  rowIndex,
-  status,
-  target,
-  actions,
-  details,
-  instance,
-}: RowPropTypes) => {
-  const { analytics, auth } = useChrome();
-  const { userData } = useGetUser(auth);
-  const isOnPremise = useAppSelector(selectIsOnPremise);
-  const isMachinesAvailable = useCockpitMachinesAvailable();
-
-  const [isExpanded, setIsExpanded] = useState(false);
-  const handleToggle = () => setIsExpanded(!isExpanded);
-  const dispatch = useDispatch();
-  const selectedBlueprintId = useAppSelector(selectSelectedBlueprintId);
-  const bootcCompose = hasBootcRequest(compose) ? compose : undefined;
-
-  const { data } = useGetComposeStatusQuery({
-    composeId: compose.id,
-  });
-
-  const options = data?.image_status.upload_status?.options as unknown as
-    | LocalUploadStatus
-    | undefined;
-
-  const handleClick = ({
-    blueprintId,
-  }: {
-    blueprintId: BlueprintItem['id'];
-  }) => {
-    if (blueprintId) {
-      dispatch(setBlueprintId(blueprintId));
-    }
-  };
-
-  return (
-    <Tbody key={compose.id} isExpanded={isExpanded}>
-      <Tr className='no-bottom-border'>
-        <Td
-          expand={{
-            rowIndex: rowIndex,
-            isExpanded: isExpanded,
-            onToggle: () => handleToggle(),
-          }}
-        />
-        <Td dataLabel='Image name'>
-          {compose.blueprint_id && !selectedBlueprintId ? (
-            <Button
-              component='a'
-              variant='link'
-              isInline
-              onClick={() =>
-                compose.blueprint_id &&
-                handleClick({ blueprintId: compose.blueprint_id })
-              }
-            >
-              {compose.image_name || compose.id}
-            </Button>
-          ) : (
-            <span>{compose.image_name || compose.id}</span>
-          )}{' '}
-          {compose.blueprint_id &&
-            (bootcCompose?.request.bootc ? (
-              <Label isCompact color='yellow'>
-                Image mode
-              </Label>
-            ) : (
-              <Label isCompact color='teal'>
-                Package mode
-              </Label>
-            ))}
-        </Td>
-        <Td
-          dataLabel='Created'
-          title={timestampToDisplayStringDetailed(compose.created_at)}
-        >
-          {timestampToDisplayString(compose.created_at)}
-        </Td>
-        <Td dataLabel='Release'>
-          {compose.request.distribution ? (
-            <Release release={compose.request.distribution} />
-          ) : (
-            <p>
-              {bootcCompose?.request.bootc?.reference &&
-                bootcReferenceToOSDisplayLabel(
-                  bootcCompose.request.bootc.reference,
-                )}
-            </p>
-          )}
-        </Td>
-        <Td dataLabel='Target'>
-          {target ? target : <Target compose={compose} />}
-        </Td>
-        {!isOnPremise && (
-          <Td dataLabel='Version'>
-            <Badge isRead>{compose.blueprint_version || 'N/A'}</Badge>
-          </Td>
-        )}
-        <Td dataLabel='Status'>{status}</Td>
-        <Td dataLabel='Instance'>{instance}</Td>
-        <Td>
-          {actions ? (
-            actions
-          ) : (
-            <ActionsColumn
-              items={defaultActions(
-                compose,
-                analytics,
-                userData?.identity.internal?.account_id,
-                isOnPremise,
-                options,
-                isMachinesAvailable,
-              )}
-            />
-          )}
-        </Td>
-      </Tr>
-      <Tr isExpanded={isExpanded}>
-        <Td colSpan={8}>
-          <ExpandableRowContent>{details}</ExpandableRowContent>
-        </Td>
-      </Tr>
-    </Tbody>
-  );
-};
-
-const defaultActions = (
-  compose: ComposesResponseItem,
-  analytics: Analytics,
-  account_id: string | undefined,
-  isOnPremise: boolean,
-  options: LocalUploadStatus | undefined,
-  isMachinesAvailable: boolean,
-) => {
-  const name = `request-${compose.id}.json`;
-
-  const handleDownload = (e: React.MouseEvent) => {
-    e.preventDefault();
-
-    const data = JSON.stringify(compose.request, null, 2);
-    const blob = new Blob([data], { type: 'application/json' });
-
-    const url = URL.createObjectURL(blob);
-    // In cockpit we're running in an iframe, the current content-security policy
-    // (set in cockpit/public/manifest.json) only allows resources from the same origin as the
-    // document (which is unique to the iframe). So create the element in the parent document.
-    const link = isOnPremise
-      ? window.parent.document.createElement('a')
-      : document.createElement('a');
-    link.href = url;
-    link.download = name;
-    link.click();
-    setTimeout(() => URL.revokeObjectURL(url), 0);
-
-    if (!isOnPremise) {
-      analytics.track(`${AMPLITUDE_MODULE_NAME} - File Downloaded`, {
-        module: AMPLITUDE_MODULE_NAME,
-        link_name: name,
-        current_path: window.location.pathname,
-        account_id: account_id || 'Not found',
-      });
-    }
-  };
-
-  const actions = [
-    {
-      title: 'Download compose request (.json)',
-      onClick: handleDownload,
-    },
-  ];
-
-  if (isOnPremise && isMachinesAvailable && options?.artifact_path) {
-    const parsedPath = path.parse(options.artifact_path);
-    const fileBrowserHref =
-      '/files#/?path=' + encodeURIComponent(parsedPath.dir);
-
-    const handleOpenInFileBrowser = (e: React.MouseEvent) => {
-      e.preventDefault();
-      cockpit.jump(fileBrowserHref, cockpit.transport.host);
-    };
-
-    actions.push({
-      title: 'Open in file browser',
-      onClick: handleOpenInFileBrowser,
-    });
-  }
-
-  return actions;
 };
 
 export default ImagesTable;
